added more error checking

+++ orgeefiles.rb
@@ -20,6 +20,11 @@ end
 config = YAML.load_file('config.yml')
 
 src_dir = config['look']['dir']
+
+unless File.directory?(src_dir)
+  raise "ERROR: '#{src_dir}' does not exist."
+end
+
 regex = Regexp.new(config['look']['regex'], Regexp::IGNORECASE)
 
 puts "- Looking for files to move in #{src_dir}"
@@ -38,8 +43,9 @@ files.each do |file|
     regex = Regexp.new(mapping['regex'], Regexp::IGNORECASE)
     if m = regex.match(file)
       dest_dir = "#{dir}/Season #{m[1].to_i}/"    # NOTE: Change this to suit
-      puts "- Attempting to move #{file} to #{dest_dir}"
+      puts "+ Moving #{file} to #{dest_dir}"
       if ARGV.flags.forreal
+        raise "Directory '#{dest_dir}' does not exist" unless File.directory?(File.dirname(dest_dir))
         Dir.mkdir(dest_dir) unless File.directory?(dest_dir)
         FileUtils.mv("#{file}", "#{dest_dir}/")
       end
